mkvmuxerutil: Fix MSVC build.
Avoid a couple implicit casts.

+++ mkvmuxer/mkvmuxerutil.cc
@@ -251,10 +251,11 @@ uint64_t EbmlElementSize(uint64_t type, uint64_t value) {
 
 uint64_t EbmlElementSize(uint64_t type, uint64_t value, uint64_t fixed_size) {
   // Size of EBML ID
-  int32_t ebml_size = GetUIntSize(type);
+  uint64_t ebml_size = static_cast<uint64_t>(GetUIntSize(type));
 
   // Datasize
-  ebml_size += (fixed_size > 0) ? fixed_size : GetUIntSize(value);
+  ebml_size +=
+      (fixed_size > 0) ? fixed_size : static_cast<uint64_t>(GetUIntSize(value));
 
   // Size of Datasize
   ebml_size++;
